removed warning "core.cpp:3090: warning: comparison between signed and unsigned integer expressions"
git-svn-id: file:///.data/svn_convert/svn@3519 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/_core/core.cpp
@@ -3083,11 +3083,11 @@ bool CorePlugin::init(bool bInit)
 
     m_main = new MainWindow;
     m_view = new UserView;
-    if ((data.geometry[WIDTH].value == -1) && (data.geometry[HEIGHT].value == -1)){
+    if ((data.geometry[WIDTH].value == (unsigned long)-1) && (data.geometry[HEIGHT].value == (unsigned long)-1)){
         data.geometry[HEIGHT].value = QApplication::desktop()->height() * 2 / 3;
         data.geometry[WIDTH].value  = data.geometry[HEIGHT].value / 3;
     }
-    if ((data.geometry[LEFT].value == -1) && (data.geometry[TOP].value == -1)){
+    if ((data.geometry[LEFT].value == (unsigned long)-1) && (data.geometry[TOP].value == (unsigned long)-1)){
         data.geometry[LEFT].value = QApplication::desktop()->width() - 25 - data.geometry[WIDTH].value;
         data.geometry[TOP].value = 5;
     }
